requirements.txt
pywebview>=3.7
pyinstaller>=5.15